refactored CSS (BEM and SMACSS)
