+++ src/components/layout/index.js
@@ -1,8 +1,6 @@
 import React from "react"
 import PropTypes from "prop-types"
 
-import "./layout.scss"
-
 
 const Layout = ({ children }) => {
   return (
