--- src/components/layout/layout.scss
@@ -1,46 +1,0 @@
-
-@import 'src/style/normalize'; // Global styles could be included in gatsby-browser.js instead of here
-main{
-}
-.headroom-wrapper{
-  position: absolute;;
-  bottom:-30px;
-  z-index: 400;
-}
-header{
-  .headroom--top{
-    position:relative;
-  }
-    background: #ffffff63;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-around;
-    .nav-item{
-                 flex-grow:1;
-
-      a{
-         .nav-link{
-          text-decoration: none;
-          transition-duration: 300ms;
-          transition-property: background-color;
-          background:black;
-          color:white;
-          padding:10px; 
-        }
-        &:hover{
-          .nav-link{
-            background:blue;
-          }
-        }
-        &:visited{
-          color:inherit;
-        }
-      }
-    }
-}
-
-@media only screen and (max-width: 600px) {
-  #main-content{
-    margin: 0;
-  }
-}
