--- src/components/portfolio/animation.scss
@@ -1,65 +1,0 @@
-
-#curve{
-    bottom: 0;
-        line-height:0;
-  position:absolute;
-    path{
-    stroke: none;
-    stroke-width: 2px;
-  }
-  svg{
-    background:none;
-  }
-  }
-  #icon-animation{
-    z-index: 10;
-    position:relative;
-    .curve{
-      position:absolute
-    }
-    .animated-icon{
-  
-  svg{
-    transform: rotate(250deg);
-  box-sizing: border-box;
-    transition-duration: 300ms;
-    transition-property:all;
-          padding:-5px;
-  overflow:visible;
-      cursor:pointer;
-  
-  
-    //background:white;
-    path{
-      transform:scale(1);
-      transform-origin: center;
-      transition-duration: 300ms;
-      transition-property:all;
-  
-      //fill:unset;
-    }
-  }
-  
-  height:20px;
-  width:20px;
-    border-radius: 5px;
-    text-align: center;
-    position:absolute;
-    top:10px;
-    &:hover{
-  border-radius: 5px;
-    svg{
-  path{
-    transform:scale(1.3);
-  
-  }      opacity: 1;
-      }
-  
-    }
-  }
-  }
-  
-  
-  
-  
-
