+++ src/components/portfolio/animation/animation.scss
@@ -1,0 +1,62 @@
+@import "../../../style/base/variables.scss";
+.animation {
+  height: 50%;
+  flex-grow: 1;
+  flex-basis: 50%;
+  background: $primary-color2;
+  position: relative;
+  width: 100%;
+  overflow: visible;
+  z-index: 200;
+  .curve {
+    bottom: 0;
+    line-height: 0;
+    position: absolute;
+    path {
+      stroke: none;
+      stroke-width: 2px;
+    }
+    svg {
+      background: none;
+    }
+  }
+}
+
+#icon-animation {
+  .animated-icon {
+    height: 20px;
+    width: 20px;
+    border-radius: 5px;
+    text-align: center;
+    position: absolute;
+    top: 10px;
+    svg {
+      transform: rotate(250deg);
+      box-sizing: border-box;
+      transition-duration: 300ms;
+      transition-property: all;
+      padding: -5px;
+      overflow: visible;
+      cursor: pointer;
+
+      //background:white;
+      path {
+        transform: scale(1);
+        transform-origin: center;
+        transition-duration: 300ms;
+        transition-property: all;
+
+        //fill:unset;
+      }
+    }
+    &:hover {
+      border-radius: 5px;
+      svg {
+        path {
+          transform: scale(1.3);
+        }
+        opacity: 1;
+      }
+    }
+  }
+}
