--- src/components/portfolio/portfolio.scss
@@ -1,108 +1,0 @@
-.stretch{
-  display: flex;
-    flex-direction: column;
-    height: 100vh;
-}
-#portfolio-page section{
-    &:first-child{
-      background:white;
-      padding:0;
-    }
-  }
-#about-section{
-  height: 100vh;
-  #intro{
-    position:relative;
-    z-index: 30;
-    display:flex;
-    align-items: center;
-    h3{
-      padding:20px;
-      color:#ffefeb;
-      border-left:2px solid black;
-
-      text-align:left;
-    }
-  }
-  h2{
-    padding:20px;
-    color: #ffa370;
-    font-size: 4em;
-    text-shadow: 2px 2px 2px black;
-    background-clip: text;
-    background-image: linear-gradient(45deg, #B2A2CD 25%, #5F5370 25%, #5F5370 50%, #B2A2CD 50%, #B2A2CD 75%, #5F5370 75%, #5F5370 100%);
-   
-
-  }
-  h3{
-    color:white;
-  }
-
-}
-
-#projects-section{
-  background: black;
-    padding: 50px;
-  .note{
-    height: 100vh;
-    border-radius: 10px;
-    margin: 50px;
-    margin-top: 0;
-    background:white;
-    border:1px solid gray;
-padding:10px;
-box-shadow: 3px 3px 5px #0000003b;
-  }
-  .project{
-    height:2000px;
-    background:black;
-    padding:10px;
-    margin:10px
-  }
-}
-#fixedBackground{
-      background-size: auto;
-    width: 100%;
-    height:100%;
-    position: absolute;
-    z-index:-100;
-    background: repeating-linear-gradient(34deg, rgb(129, 208, 248) 0px, rgb(102, 202, 244) 60px, rgb(0, 0, 0) 10px);
-    &.type2{
-      background:repeating-linear-gradient(46deg, #0aa3c8 0px, #64ceda 3px, #0a8ec2 7px, black 3px);         
-     }
-         &.type3{
-          background: repeating-linear-gradient(40deg, #73caf5 0px, #66caf482 60px, #66caf4 130px, black 10px);
-     }
-  }
-#fixedBackground2{
-  position: absolute;
-
-      background-size: auto;
-    width: 100%;
-    height:100%;
-    z-index:-100;
-    //background: repeating-linear-gradient(-45deg, #000000, #69369e85 2px, #1600189e 60px, #00000075 10px);        /* background-image: url(https://yourtechclass.com/wp-content/uploads/2018/01/yy3.png); */
-}
-
-#fixedBackground3{
-  background-size: auto;
-width: 100%;
-height:100%;
-position: absolute;
-z-index:-100;
-background: linear-gradient(to bottom, rgba(3, 123, 157, 0.26) 0%, #02638300 100%);
-    /* background-image: url(https://yourtechclass.com/wp-content/uploads/2018/01/yy3.png); */
- &.type2{
-  background:linear-gradient(#059abb,rgba(60,172,204,.46) 30%,#8ae0f0 90%);
- }
- &.type3{
-  background: linear-gradient(to top, rgb(0, 141, 185) 0%, #00475f0d 100%);
-}
-  }
-
-@media only screen and (max-width: 600px) {
-  #about-section #intro{
-    margin: 0;
-    display:inline-block;
-  }
-}
