+++ src/style/base/normalize.scss
@@ -1,3 +1,4 @@
+
 html {
     font-family: sans-serif;
     -ms-text-size-adjust: 100%;
