+++ src/style/base/reset.scss
@@ -1,0 +1,10 @@
+h1,h2,h3,h4,h5,h6{
+    margin:0;
+    padding:0;
+}
+html{
+    box-sizing:border-box;
+}
+a:visited {
+    color: inherit;
+  }
