+++ src/style/base/variables.scss
@@ -1,0 +1,4 @@
+$primary-color1:#099cbd;
+$primary-color2:#8ae0f0;
+
+$secondary-color1:#ffa370;
