+++ src/style/layout/layout.scss
@@ -1,0 +1,30 @@
+.headroom-wrapper {
+  z-index: 400;
+}
+.headroom--top {
+  position: relative;
+}
+header {
+  background: #ffffff63;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+}
+.inner-container {
+  padding: 20px;
+}
+#main-content {
+  padding: 0 9%;
+  background: linear-gradient(
+    #059abb,
+    rgba(60, 172, 204, 0.46) 30%,
+    #8ae0f0 90%
+  );
+}
+@media only screen and (max-width: 800px) {
+  #main-content {
+    padding: 0;
+    background: none;
+    margin: 0;
+  }
+}
